pyproject.toml
[project]
name = "ya-trash-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fire>=0.7.0",
    "llama-stack-client>=0.2.10",
    "openai-agents>=0.0.18",
    "requests>=2.32.4",
    "streamlit>=1.45.1",
]
